Move all operations into folder
