+++ src/utils/github/index.ts
@@ -1,5 +1,10 @@
 export { processGitHubContent } from './content-processor';
 export { ghFetch, fetchRepo, fetchDirectory, fetchFile } from './api-client';
-export { processRepo, processPath, processFile } from './repo-operations';
-export { processPullRequest } from './pr-operations';
-export { processIssue } from './issue-operations';
+export {
+  processRepo,
+  processPath,
+  processFile,
+  processDirectoryContents,
+  processPullRequest,
+  processIssue
+} from './operations';
