+++ src/utils/github/operations/index.ts
@@ -2,3 +2,5 @@ export { processRepo } from './repo';
 export { processPath } from './path';
 export { processFile } from './file';
 export { processDirectoryContents } from './directory';
+export { processPullRequest } from './pull-request';
+export { processIssue } from './issue';
